chore: moved base test setup into lib + configurable pc
