+++ settings.gradle.kts
@@ -301,6 +301,7 @@ include(":system-tests:management-api:management-api-test-runner")
 include(":system-tests:management-api:management-api-test-runtime")
 include(":system-tests:protocol-2025-test")
 include(":system-tests:protocol-tck:tck-extension")
+include(":system-tests:protocol-tck:lib:tck-lib")
 include(":system-tests:telemetry:telemetry-test-runner")
 include(":system-tests:telemetry:telemetry-test-runtime")
 include(":system-tests:version-api:version-api-test-runner")
