+++ system-tests/protocol-tck/lib/tck-lib/src/main/java/org/eclipse/edc/tck/dsp/guard/StatefulEntityTriggerRegistry.java
