+++ system-tests/protocol-tck/lib/tck-lib/src/test/java/org/eclipse/edc/tck/dsp/recorder/StepRecorderTest.java
