+++ system-tests/protocol-tck/tck-extension/src/main/java/org/eclipse/edc/tck/dsp/guard/GuardAssembly.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2025 Metaform Systems, Inc.
+ *  Copyright (c) 2026 Metaform Systems, Inc.
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -12,95 +12,33 @@
  *
  */
 
-package org.eclipse.edc.tck.dsp.data;
+package org.eclipse.edc.tck.dsp.guard;
 
-import org.eclipse.edc.connector.controlplane.asset.spi.domain.Asset;
 import org.eclipse.edc.connector.controlplane.contract.spi.event.contractnegotiation.ContractNegotiationAccepted;
 import org.eclipse.edc.connector.controlplane.contract.spi.event.contractnegotiation.ContractNegotiationAgreed;
 import org.eclipse.edc.connector.controlplane.contract.spi.event.contractnegotiation.ContractNegotiationEvent;
 import org.eclipse.edc.connector.controlplane.contract.spi.event.contractnegotiation.ContractNegotiationOffered;
 import org.eclipse.edc.connector.controlplane.contract.spi.event.contractnegotiation.ContractNegotiationRequested;
-import org.eclipse.edc.connector.controlplane.contract.spi.types.agreement.ContractAgreement;
 import org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractNegotiation;
-import org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractNegotiationStates;
-import org.eclipse.edc.connector.controlplane.contract.spi.types.offer.ContractDefinition;
-import org.eclipse.edc.connector.controlplane.policy.spi.PolicyDefinition;
 import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessEvent;
 import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessInitiated;
 import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessRequested;
 import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessStarted;
 import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessSuspended;
 import org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferProcess;
-import org.eclipse.edc.policy.model.Action;
-import org.eclipse.edc.policy.model.Permission;
-import org.eclipse.edc.policy.model.Policy;
-import org.eclipse.edc.spi.types.domain.DataAddress;
-import org.eclipse.edc.tck.dsp.guard.Trigger;
 import org.eclipse.edc.tck.dsp.recorder.StepRecorder;
 
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 
-import static java.util.stream.Collectors.toSet;
 import static org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractNegotiationStates.REQUESTED;
 
 /**
- * Assembles data for the TCK scenarios.
+ * Assembles guard for the TCK scenarios.
  */
-public class DataAssembly {
-    public static final String PARTICIPANT_CONTEXT_ID = "participantContextId";
-    private static final Set<String> ASSET_IDS = Set.of("ACN0101", "ACN0102", "ACN0103", "ACN0104",
-            "ACN0201", "ACN0202", "ACN0203", "ACN0204", "ACN0205", "ACN0206", "ACN0207",
-            "ACN0301", "ACN0302", "ACN0303", "ACN0304",
-            "CAT0101", "CAT0102");
-    private static final Set<String> AGREEMENT_IDS = Set.of(
-            "ATP0101", "ATP0102", "ATP0103", "ATP0104", "ATP0105",
-            "ATP0201", "ATP0202", "ATP0203", "ATP0204", "ATP0205",
-            "ATP0301", "ATP0302", "ATP0303", "ATP0304", "ATP0305", "ATP0306",
-            "ATPC0101", "ATPC0102", "ATPC0103", "ATPC0104", "ATPC0105",
-            "ATPC0201", "ATPC0202", "ATPC0203", "ATPC0204", "ATPC0205",
-            "ATPC0301", "ATPC0302", "ATPC0303", "ATPC0304", "ATPC0305", "ATPC0306");
-    private static final String POLICY_ID = "P123";
-    private static final String CONTRACT_DEFINITION_ID = "CD123";
-
-    private DataAssembly() {
-    }
-
-    public static Set<Asset> createAssets() {
-        var assets = ASSET_IDS.stream().map(DataAssembly::createAsset).collect(toSet());
-
-        assets.add(Asset.Builder.newInstance().id("ATP0101")
-                .dataAddress(DataAddress.Builder.newInstance().type("HttpData").build())
-                .participantContextId(PARTICIPANT_CONTEXT_ID)
-                .build());
-        return assets;
-    }
-
-    public static Set<PolicyDefinition> createPolicyDefinitions() {
-        var permission = Permission.Builder.newInstance()
-                .action(Action.Builder.newInstance().type("http://www.w3.org/ns/odrl/2/use").build())
-                .build();
-        return Set.of(PolicyDefinition.Builder.newInstance()
-                .id(POLICY_ID)
-                .policy(Policy.Builder.newInstance().permission(permission).build())
-                .participantContextId(PARTICIPANT_CONTEXT_ID)
-                .build());
-    }
-
-    public static Set<ContractDefinition> createContractDefinitions() {
-        return Set.of(ContractDefinition.Builder.newInstance()
-                .id(CONTRACT_DEFINITION_ID)
-                .accessPolicyId(POLICY_ID)
-                .contractPolicyId(POLICY_ID)
-                .participantContextId(PARTICIPANT_CONTEXT_ID)
-                .build());
-    }
-
-
-    public static Set<ContractNegotiation> createContractNegotiations() {
-        return AGREEMENT_IDS.stream().map(DataAssembly::createContractNegotiation).collect(toSet());
+public class GuardAssembly {
+    private GuardAssembly() {
     }
 
     public static StepRecorder<ContractNegotiation> createNegotiationRecorder() {
@@ -158,10 +96,10 @@ public class DataAssembly {
 
         recorder.record("ACN0302", ContractNegotiation::transitionOffering);
         recorder.record("ACN0303", ContractNegotiation::transitionOffering)
-                .record("ACN0303", DataAssembly::noop);
+                .record("ACN0303", GuardAssembly::noop);
 
         recorder.record("ACN0304", ContractNegotiation::transitionOffering)
-                .record("ACN0304", DataAssembly::noop);
+                .record("ACN0304", GuardAssembly::noop);
 
     }
 
@@ -340,35 +278,6 @@ public class DataAssembly {
         }, action);
     }
 
-    private static Asset createAsset(String id) {
-        return Asset.Builder.newInstance()
-                .id(id)
-                .dataAddress(DataAddress.Builder.newInstance().type("HttpData").build())
-                .participantContextId(PARTICIPANT_CONTEXT_ID)
-                .build();
-    }
-
-    private static ContractNegotiation createContractNegotiation(String id) {
-
-        return ContractNegotiation.Builder.newInstance()
-                .contractAgreement(ContractAgreement.Builder.newInstance()
-                        .id(id)
-                        .providerId("providerId")
-                        .consumerId("TCK_PARTICIPANT")
-                        .assetId("ATP0101")
-                        .contractSigningDate(System.currentTimeMillis())
-                        .policy(Policy.Builder.newInstance().build())
-                        .participantContextId(PARTICIPANT_CONTEXT_ID)
-                        .agreementId(id)
-                        .build())
-                .type(ContractNegotiation.Type.PROVIDER)
-                .state(ContractNegotiationStates.FINALIZED.code())
-                .counterPartyId("counterPartyId")
-                .counterPartyAddress("https://test.com")
-                .protocol("test")
-                .participantContextId(PARTICIPANT_CONTEXT_ID)
-                .build();
-    }
 
     private static void noop(ContractNegotiation cn) {
